mx7dvar_som: Remove I2C3 configuration and initialization
Remove the definition and initialization of the I2C3 interface from
the mx7dvar_som. These pins are for the CAN2 interface and should
no longer be used as I2C. Delete the `i2c_pad_info3` structure. And,
remove the `setup_i2c` call for I2C3 in `board_early_init_f`.

Signed-off-by: Diego Dorta <diego.d@variscite.com>